+++ planning/templates.js
@@ -84,7 +84,7 @@ angular.module('90Tech.planning').run(['$templateCache', function($templateCache
     "\n" +
     "    </div>\n" +
     "    <div class=\"event\" style=\"display: inline-block; position: relative; height: 100%;\" data-context-menu=\"planning/templates/planning-context-menu.html\"\n" +
-    "         tooltip-append-to-body=\"true\" tooltip-placement=\"{{event.tooltipTemplate? 'right': 'top'}}\"\n" +
+    "         tooltip-append-to-body=\"true\" tooltip-placement=\"{{event.tooltipTemplate? 'auto': 'top'}}\"\n" +
     "         ng-attr-uib-tooltip=\"{{!event.tooltipTemplate ? event.tooltip : undefined}}\"\n" +
     "         ng-attr-uib-tooltip-template=\"{{'' + event.tooltipTemplate}}\" tooltip-class=\"planning-event-tooltip\"\n" +
     "         ng-style=\"{\n" +
@@ -179,7 +179,7 @@ angular.module('90Tech.planning').run(['$templateCache', function($templateCache
     "        </div>\n" +
     "\n" +
     "        <div class=\"event\" style=\"display: inline-block; position: relative; width: 100%;\"\n" +
-    "             tooltip-append-to-body=\"true\" tooltip-placement=\"{{event.tooltipTemplate? 'right': 'top'}}\"\n" +
+    "             tooltip-append-to-body=\"true\" tooltip-placement=\"{{event.tooltipTemplate? 'auto': 'top'}}\"\n" +
     "             ng-attr-uib-tooltip=\"{{!event.tooltipTemplate ? event.tooltip : undefined}}\"\n" +
     "             ng-attr-uib-tooltip-template=\"{{'' + event.tooltipTemplate}}\" tooltip-class=\"planning-event-tooltip\"\n" +
     "             data-context-menu=\"planning/templates/planning-context-menu.html\"\n" +
@@ -235,7 +235,7 @@ angular.module('90Tech.planning').run(['$templateCache', function($templateCache
     "                ng-style=\"{'top': ev.style.top, 'height': ev.style.height, 'background-color': ev.style['background-color'], 'color' : ev.style.color, 'width': ev.style.width, 'left': ev.style.left}\"\n" +
     "                style=\"position:absolute;border: 1px solid black;pointer-events: auto; overflow: hidden;\"\n" +
     "                ng-click=\"planning.weekEventCallback({event: ev})\"\n" +
-    "                tooltip-append-to-body=\"true\" tooltip-placement=\"{{ev.tooltipTemplate? 'right': 'top'}}\"\n" +
+    "                tooltip-append-to-body=\"true\" tooltip-placement=\"{{ev.tooltipTemplate? 'auto': 'top'}}\"\n" +
     "                ng-attr-uib-tooltip=\"{{!ev.tooltipTemplate ? ev.tooltip : undefined}}\"\n" +
     "                ng-attr-uib-tooltip-template=\"{{'' + ev.tooltipTemplate}}\" tooltip-class=\"planning-event-tooltip\"\n" +
     "                class=\"single-day-event\">\n" +
@@ -253,7 +253,7 @@ angular.module('90Tech.planning').run(['$templateCache', function($templateCache
     "             zl-planning-drag-drop\n" +
     "             zl-drag=\"event\"\n" +
     "             ng-click=\"planning.weekEventCallback({event: event})\"\n" +
-    "             tooltip-append-to-body=\"true\" tooltip-placement=\"{{event.tooltipTemplate? 'right': 'top'}}\"\n" +
+    "             tooltip-append-to-body=\"true\" tooltip-placement=\"{{event.tooltipTemplate? 'auto': 'top'}}\"\n" +
     "             ng-attr-uib-tooltip=\"{{!event.tooltipTemplate ? event.tooltip : undefined}}\"\n" +
     "             ng-attr-uib-tooltip-template=\"{{'' + event.tooltipTemplate}}\" tooltip-class=\"planning-event-tooltip\"\n" +
     "             ng-style=\"{'top': event.style.top, 'height': event.style.height, 'background-color': event.style['background-color'], 'color' : event.style.color, 'width': event.style.width, 'left': event.style.left}\">\n" +