+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version numbers correspond to `bower.json` version
 # Features
 
 - Update AngularJS to 1.7.0
+- Switch to position: auto for custom event tooltip
 
 # 1.5.4